+++ quic/handshake/TransportParameters.h
@@ -8,7 +8,6 @@
 
 #pragma once
 
-#include <fizz/record/Extensions.h>
 #include <quic/QuicConstants.h>
 #include <quic/QuicException.h>
 #include <quic/codec/Types.h>
@@ -142,146 +141,4 @@ inline TransportParameter encodeStatelessResetToken(
   statelessReset.value = folly::IOBuf::copyBuffer(token.data(), token.size());
   return statelessReset;
 }
-
-inline fizz::Extension encodeExtension(
-    const quic::ClientTransportParameters& params) {
-  fizz::Extension ext;
-  ext.extension_type = fizz::ExtensionType::quic_transport_parameters;
-  ext.extension_data = folly::IOBuf::create(0);
-  folly::io::Appender appender(ext.extension_data.get(), 40);
-  if (params.initial_version) {
-    fizz::detail::write(params.initial_version.value(), appender);
-  }
-  fizz::detail::writeVector<uint16_t>(params.parameters, appender);
-  return ext;
-}
-
-inline fizz::Extension encodeExtension(
-    const quic::ServerTransportParameters& params) {
-  fizz::Extension ext;
-  ext.extension_type = fizz::ExtensionType::quic_transport_parameters;
-  ext.extension_data = folly::IOBuf::create(0);
-  folly::io::Appender appender(ext.extension_data.get(), 40);
-  if (params.negotiated_version) {
-    fizz::detail::write(params.negotiated_version.value(), appender);
-    fizz::detail::writeVector<uint8_t>(params.supported_versions, appender);
-  }
-  fizz::detail::writeVector<uint16_t>(params.parameters, appender);
-  return ext;
-}
-
-inline fizz::Extension encodeExtension(
-    const quic::TicketTransportParameters& params) {
-  fizz::Extension ext;
-  ext.extension_type = fizz::ExtensionType::quic_transport_parameters;
-  ext.extension_data = folly::IOBuf::create(0);
-  folly::io::Appender appender(ext.extension_data.get(), 40);
-  if (params.negotiated_version) {
-    fizz::detail::write(params.negotiated_version.value(), appender);
-  }
-  fizz::detail::writeVector<uint16_t>(params.parameters, appender);
-  return ext;
-}
-
 } // namespace quic
-
-namespace fizz {
-
-template <>
-inline folly::Optional<quic::ClientTransportParameters> getExtension(
-    const std::vector<Extension>& extensions) {
-  auto it = findExtension(extensions, ExtensionType::quic_transport_parameters);
-  if (it == extensions.end()) {
-    return folly::none;
-  }
-  quic::ClientTransportParameters parameters;
-  folly::io::Cursor cursor(it->extension_data.get());
-  folly::io::Cursor versionCursor(it->extension_data.get());
-  quic::QuicVersion tryVersion{quic::QuicVersion::MVFST_INVALID};
-  if (cursor.canAdvance(sizeof(tryVersion))) {
-    detail::read(tryVersion, versionCursor);
-  }
-  if (tryVersion == quic::QuicVersion::MVFST_OLD) {
-    parameters.initial_version = tryVersion;
-    cursor.skip(sizeof(tryVersion));
-  }
-  detail::readVector<uint16_t>(parameters.parameters, cursor);
-  return parameters;
-}
-
-template <>
-inline folly::Optional<quic::ServerTransportParameters> getExtension(
-    const std::vector<Extension>& extensions) {
-  auto it = findExtension(extensions, ExtensionType::quic_transport_parameters);
-  if (it == extensions.end()) {
-    return folly::none;
-  }
-  quic::ServerTransportParameters parameters;
-  folly::io::Cursor cursor(it->extension_data.get());
-  folly::io::Cursor versionCursor(it->extension_data.get());
-  quic::QuicVersion tryVersion{quic::QuicVersion::MVFST_INVALID};
-  if (cursor.canAdvance(sizeof(tryVersion))) {
-    detail::read(tryVersion, versionCursor);
-  }
-  if (tryVersion == quic::QuicVersion::MVFST_OLD) {
-    parameters.negotiated_version = tryVersion;
-    cursor.skip(sizeof(tryVersion));
-    detail::readVector<uint8_t>(parameters.supported_versions, cursor);
-  }
-  detail::readVector<uint16_t>(parameters.parameters, cursor);
-  return parameters;
-}
-
-template <>
-inline folly::Optional<quic::TicketTransportParameters> getExtension(
-    const std::vector<Extension>& extensions) {
-  auto it = findExtension(extensions, ExtensionType::quic_transport_parameters);
-  if (it == extensions.end()) {
-    return folly::none;
-  }
-  quic::TicketTransportParameters parameters;
-  folly::io::Cursor cursor(it->extension_data.get());
-  folly::io::Cursor versionCursor(it->extension_data.get());
-  quic::QuicVersion tryVersion{quic::QuicVersion::MVFST_INVALID};
-  if (versionCursor.canAdvance(sizeof(tryVersion))) {
-    detail::read(tryVersion, versionCursor);
-  }
-  if (tryVersion == quic::QuicVersion::MVFST_OLD) {
-    parameters.negotiated_version = tryVersion;
-    cursor.skip(sizeof(tryVersion));
-  }
-  detail::readVector<uint16_t>(parameters.parameters, cursor);
-  return parameters;
-}
-
-namespace detail {
-
-template <>
-struct Reader<quic::TransportParameter> {
-  template <class T>
-  size_t read(quic::TransportParameter& param, folly::io::Cursor& cursor) {
-    size_t len = 0;
-    len += detail::read(param.parameter, cursor);
-    len += readBuf<uint16_t>(param.value, cursor);
-    return len;
-  }
-};
-
-template <>
-struct Writer<quic::TransportParameter> {
-  template <class T>
-  void write(const quic::TransportParameter& param, folly::io::Appender& out) {
-    detail::write(param.parameter, out);
-    detail::writeBuf<uint16_t>(param.value, out);
-  }
-};
-
-template <>
-struct Sizer<quic::TransportParameter> {
-  template <class T>
-  size_t getSize(const quic::TransportParameter& param) {
-    return sizeof(param.parameter) + getBufSize<uint16_t>(param.value);
-  }
-};
-} // namespace detail
-} // namespace fizz
